requirements.txt
# vcpmctool/requirements.txt (Nội dung mới)
flet==0.24.1
pandas==2.2.0
openpyxl==3.1.2
python-dateutil==2.9.0